requirements.txt
Django==4.1.3
djangorestframework==3.14.0
pip==22.0.4
PyJWT==2.6.0
python_bcrypt==0.3.2
